+++ packages/natives/native/index.js
@@ -24,7 +24,7 @@ export const Shell = nativeBindings.Shell;
 
 // functions
 export const __ompInstallTokioRuntime = nativeBindings.__ompInstallTokioRuntime;
-export const __piNativesV16_2_12 = nativeBindings.__piNativesV16_2_12;
+export const __piNativesV16_2_13 = nativeBindings.__piNativesV16_2_13;
 export const applyBashFixups = nativeBindings.applyBashFixups;
 export const astEdit = nativeBindings.astEdit;
 export const astGrep = nativeBindings.astGrep;